Separate Rodan methods to new file & class and instantiate in pixel.js
This process will be automated by a python script so that the PixelWrapper instantiation lines will be automatically injected into Pixel.js
